Creando Server con express - Routes de express - Consumindo modulos de Arq Hexagonal con users en memoria

+++ src/lib/User/application/UserDelete/UserDelete.ts
@@ -1,5 +1,6 @@
 import { User } from "../../domain/User";
 import { UserId } from "../../domain/UserId";
+import { UserNotFoundError } from "../../domain/UserNotFoundError";
 import { UserRepository } from "../../domain/UserRepository";
 
 export class UserDelete {
@@ -7,6 +8,9 @@ export class UserDelete {
     async run(
         id: string
     ): Promise<void> {
+        const userExist= await this.repository.getOneById(new UserId(id));
+        if(!userExist) 
+            throw new UserNotFoundError("User not found")
         await this.repository.delete(new UserId(id));
     }
 }
